+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
-.vscode/
+.vscode
+.eslintrc.js
 .env
 package-lock.json
