+++ games/structures/Matrix.js
@@ -1,0 +1,10 @@
+export class Matrix extends Array {
+    constructor(rows, cols) {
+        super(rows);
+
+        this.rows = rows;
+        this.cols = cols;
+
+        this.fill([]);
+    }
+}
